requirements.txt
matplotlib==3.5.1
mlxtend==0.19.0
statsmodels==0.13.2
flake8==4.0.1
numpy==1.22.3
pandas==1.4.2
joblib==1.1.0
jupyter==1.0.0
mlflow==1.26.1
scikit_learn==1.0
python-dotenv==0.19.2
Sphinx==5.0.2
sphinx-rtd-theme==1.0.0
shap==0.39.0 